🔥 chore(broadcast.js): remove broadcast.js file The file broadcast.js has been removed from the project. This file was previously used to handle broadcast events, but it is no longer needed and has been deleted.
🔥 chore(connect.js): remove connect.js file
The connect.js file has been removed from the project. This file was previously used to handle connections to a Lambda function, but it is no longer needed and has been deleted.

🔥 chore(default.js): remove default.js file
The file default.js has been removed from the project. This file was generated by webpack and contained a warning message about using the "eval-source-map" devtool in production. Since this file is not needed in the project, it has been removed to reduce clutter.

🔥 chore(disconnect.js): remove disconnect.js file
The disconnect.js file has been removed from the project. This file was previously used to handle disconnection events, but it is no longer needed and has been deleted.

🐛 fix(serverless.yml): exclude aws-sdk and @aws-sdk from webpack includeModules
✨ feat(serverless.yml): update runtime to nodejs14.x and add IAM role statements
The aws-sdk and @aws-sdk modules are now excluded from the webpack includeModules configuration. This is because these modules are already available in the AWS Lambda environment and do not need to be included in the deployment package. The runtime has been updated to nodejs14.x and IAM role statements have been added to allow the Lambda function to invoke other functions and invoke them asynchronously.