📝 chore(terraform): add terraform state file and environment file This commit adds the terraform state file and environment file. The environment file contains the name of the environment, which is "webrtc-signaling-server". The terraform state file contains the state of the infrastructure managed by terraform.
🚀 chore(aws-user-data.sh): add NodeJS 17 installation and Docker installation
🆕 feat(az.tf): add AWS availability zones data source
🆕 feat(db-output.tf): add output variables for DocDB endpoint and username
🆕 feat(docdb-main.tf): add variables for DocDB database name, instance class, and password
This commit adds NodeJS 17 installation and Docker installation to the aws-user-data.sh script. It also adds a new data source for AWS availability zones in az.tf. In db-output.tf, output variables for DocDB endpoint and username are added. Finally, docdb-main.tf adds variables for DocDB database name, instance class, and password.

🆕 chore(docdb-variables.tf): add Terraform configuration for AWS DocumentDB cluster
This commit adds Terraform configuration for creating an AWS DocumentDB cluster. The configuration includes a subnet group, a cluster instance, a cluster, and a cluster parameter group. The configuration is parameterized with variables such as app_name, docdb_instance_class, and docdb_password. The configuration is designed to be used in conjunction with other Terraform configurations to create a complete infrastructure stack.

🚀 feat(linux-vm-main.tf): add security group for DocumentDB
This commit adds a new security group for DocumentDB. The security group allows traffic from the aws_security_group.aws-linux-sg security group on port 3306. Additionally, it allows HTTP traffic on port 80 from any IP address within the 0.0.0.0/16 CIDR block.

🔧 chore(network-main.tf): refactor subnet resource to use count and add tags
This commit refactors the subnet resource to use count instead of duplicating the resource block. The count is determined by the number of public and private subnets specified in the variables. Tags are also added to the subnet resources to improve readability and organization.

🚧 chore(network-main.tf): add private route table and association
This commit adds a new resource for a private route table and its association. The private route table is associated with the private subnets in the VPC. This is a work in progress and is not yet complete.

🆕 feat(network): add my_ip variable to network-variables.tf
👷‍♂️ chore(network): add aws_instance_public_dns output to network-output.tf
The my_ip variable is added to network-variables.tf to allow for the specification of an IP address range. This variable can be used to restrict access to the network from a specific IP address range. The aws_instance_public_dns output is added to network-output.tf to allow for the retrieval of the public DNS name of the Linux server instance. This output can be used to connect to the instance via SSH.