🚀 feat(docdb-main.tf): add AWS DocumentDB cluster and instance resources This commit adds AWS DocumentDB cluster and instance resources to the Terraform configuration file. The `aws_docdb_subnet_group` resource creates a subnet group for the DocumentDB cluster. The `aws_docdb_cluster_instance` resource creates a DocumentDB instance with the specified instance class. The `aws_docdb_cluster` resource creates a DocumentDB cluster with the specified configuration, including the subnet group, security group, and cluster parameter group. The `aws_docdb_cluster_parameter_group` resource creates a parameter group for the DocumentDB cluster with the specified configuration.
🔧 chore(infra): refactor docdb-variables.tf and network-main.tf
The docdb-variables.tf file has been refactored to use variables for the db_name, docdb_instance_class, and docdb_password. This makes it easier to manage the values of these variables and allows for more flexibility in the future.

The network-main.tf file has been reformatted for better readability. The vpc_id and cidr_block attributes of the aws_subnet.private-subnet resource have been aligned for better readability.